further quickcheck ...
